# internal/server/handlers/program_handler.go
package handlers

import (
	"github.com/gin-gonic/gin"

	"s-portal/internal/domain/model"
	"s-portal/internal/domain/service"
)

type ProgramHandler struct {
	programService *service.ProgramService
}

func NewProgramHandler(programService *service.ProgramService) *ProgramHandler {
	return &ProgramHandler{
		programService: programService,
	}
}

//	@Summary		GetAllPrograms
//	@Description	get all programs
//	@Tags			Program
//	@Accept			json
//	@Produce		json
//	@Success		200	{array}		model.Program
//	@Failure		404	{object}	string
//	@Router			/program [GET]
func (h *ProgramHandler) GetAllPrograms(context *gin.Context) {
	programs, err := h.programService.GetAllPrograms()
	if err != nil {
		// Handle error
		context.JSON(404, gin.H{
			"message": "No programs found",
		})
	}
	// Return programs
	context.JSON(200, gin.H{
		"message": programs,
	})
}

//	@Summary		CreateProgram
//	@Description	create a new program
//	@Tags			Program
//	@Accept			json
//	@Produce		json
//	@Param			program	body		model.Program	true	"Program object"
//	@Success		200		{object}	string
//	@Failure		400		{object}	string
//	@Router			/program [POST]
func (h *ProgramHandler) CreateProgram(context *gin.Context) {
	program := model.Program{}

	if err := context.ShouldBindJSON(&program); err != nil {
		// Handle error
		context.JSON(400, gin.H{
			"message": err.Error(),
		})
		return
	}

	if err := h.programService.CreateProgram(&program); err != nil {
		// Handle error
		context.JSON(400, gin.H{
			"message": err.Error(),
		})
		return
	}

	// Return success message
	context.JSON(200, gin.H{
		"message": "Program created successfully",
	})
}

//	@Summary		GetProgramByID
//	@Description	get a program by id
//	@Tags			Program
//	@Accept			json
//	@Produce		json
//	@Param			id	path		string	true	"program id"
//	@Success		200	{object}	model.Program
//	@Failure		404	{object}	string
//	@Router			/program/:id [GET]
func (h *ProgramHandler) GetProgramByID(context *gin.Context) {
	id := context.Param("id")
	program, err := h.programService.GetProgramByID(id)
	if err != nil {
		// Handle error
		context.JSON(404, gin.H{
			"message": err.Error(),
		})
		return;
	}
	// Return programs
	context.JSON(200, gin.H{
		"message": program,
	})
}

//	@Summary		UpdateProgramByID
//	@Description	update a program by id
//	@Tags			Program
//	@Accept			json
//	@Produce		json
//	@Param			id	path		string	true	"program id"
//	@Success		200	{object}	string
//	@Failure		404	{object}	string
//	@Router			/program/update/:id [PUT]
func (h *ProgramHandler) UpdateProgramByID(context *gin.Context) {
	id := context.Param("id")
	err := h.programService.UpdateProgramByID(context, id)
	if err != nil {
		// Handle error
		context.JSON(404, gin.H{
			"message": err.Error(),
		})
		return;
	}
	// Return success message
	context.JSON(200, gin.H{
		"message": "Program updated successfully",
	})
}

//	@Summary		DeleteProgramByID
//	@Description	delete a program by id
//	@Tags			Program
//	@Accept			json
//	@Produce		json
//	@Param			id	path		string	true	"program id"
//	@Success		200	{object}	string
//	@Failure		404	{object}	string
//	@Router			/program/delete/:id [DELETE]
func (h *ProgramHandler) DeleteProgramByID(context *gin.Context) {
	id := context.Param("id")
	err := h.programService.DeleteProgramByID(id)
	if err != nil {
		// Handle error
		context.JSON(404, gin.H{
			"message": err.Error(),
		})
		return;
	}
	// Return success message
	context.JSON(200, gin.H{
		"message": "Program deleted successfully",
	})
}